Support USB super speed in fastboot
Refine Self USB device mode:
1. USB device mode driver support 64Bit base address
2. Get xHCI base address from PCI config
Support super speed in fastboot:
1. Update needed usb3.0 descriptors to support super speed mode
   With self_usb_device_mode_protocol enabled, kernelflinger
   support usb super speed and high speed.
2. With self_usb_device_mode_protocol disabled, kernelflinger
   only support usb high speed.

Tracked-On: OAM-88234

+++ libefiusb/device_mode/XdciDevice.h
@@ -20,7 +20,7 @@
 //
 typedef struct {
   USB_CONTROLLER_ID  ControllerId; // Controller ID of the core
-  UINT32             BaseAddress;  // Base address of the controller registers and on-chip memory
+  UINT64             BaseAddress;  // Base address of the controller registers and on-chip memory
   UINT32             Flags;        // Initialization flags
   USB_SPEED          Speed;        // Desired USB bus Speed
   USB_ROLE           Role;         // Default USB role
